fix(preview): renaming display to preview
